Move orgs action to orgs file from index
